add: helm as separate plugin
